Added "method" as an alternative function keyword.
Note that maybe in the future this want to be split up into the function and the method declaration with seperate keyword names (i.e. "keyword.other.method.ph").
For now this is okay as syntactically (but not semantically) the method keyword is identical to the function keyword (and we nevertheless want to have the grammar be lax in syntax highlighting for it to not brake at the first small syntactical mistake).